+++ sdk/core/azure-core/test/ut/curl_options.cpp
@@ -44,7 +44,8 @@ namespace Azure { namespace Core { namespace Test {
     Azure::Core::Http::Request request(Azure::Core::Http::HttpMethod::Get, url);
 
     std::unique_ptr<Azure::Core::Http::RawResponse> response;
-    EXPECT_NO_THROW(response = pipeline.Send(request, Azure::Core::GetApplicationContext()));
+    EXPECT_NO_THROW(
+        response = pipeline.Send(request, Azure::Core::Context::GetApplicationContext()));
     auto responseCode = response->GetStatusCode();
     int expectedCode = 200;
     EXPECT_PRED2(
@@ -73,7 +74,8 @@ namespace Azure { namespace Core { namespace Test {
     Azure::Core::Http::Request request(Azure::Core::Http::HttpMethod::Get, url);
 
     std::unique_ptr<Azure::Core::Http::RawResponse> response;
-    EXPECT_NO_THROW(response = pipeline.Send(request, Azure::Core::GetApplicationContext()));
+    EXPECT_NO_THROW(
+        response = pipeline.Send(request, Azure::Core::Context::GetApplicationContext()));
     auto responseCode = response->GetStatusCode();
     int expectedCode = 200;
     EXPECT_PRED2(
@@ -105,10 +107,8 @@ namespace Azure { namespace Core { namespace Test {
     Azure::Core::Http::Request request(Azure::Core::Http::HttpMethod::Get, url);
 
     std::unique_ptr<Azure::Core::Http::RawResponse> response;
-    EXPECT_NO_THROW(response = pipeline.Send(Azure::Core::GetApplicationContext(), request));
-    auto responseCode = response->GetStatusCode();
-    int expectedCode = 200;
-    EXPECT_PRED2(
+    EXPECT_NO_THROW(response = pipeline.Send(Azure::Core::Context::GetApplicationContext(),
+  request)); auto responseCode = response->GetStatusCode(); int expectedCode = 200; EXPECT_PRED2(
         [](int expected, int code) { return expected == code; },
         expectedCode,
         static_cast<typename std::underlying_type<Azure::Core::Http::HttpStatusCode>::type>(
@@ -132,10 +132,8 @@ namespace Azure { namespace Core { namespace Test {
     Azure::Core::Http::Request request(Azure::Core::Http::HttpMethod::Get, url);
 
     std::unique_ptr<Azure::Core::Http::RawResponse> response;
-    EXPECT_NO_THROW(response = pipeline.Send(Azure::Core::GetApplicationContext(), request));
-    auto responseCode = response->GetStatusCode();
-    int expectedCode = 200;
-    EXPECT_PRED2(
+    EXPECT_NO_THROW(response = pipeline.Send(Azure::Core::Context::GetApplicationContext(),
+  request)); auto responseCode = response->GetStatusCode(); int expectedCode = 200; EXPECT_PRED2(
         [](int expected, int code) { return expected == code; },
         expectedCode,
         static_cast<typename std::underlying_type<Azure::Core::Http::HttpStatusCode>::type>(
@@ -159,10 +157,8 @@ namespace Azure { namespace Core { namespace Test {
     Azure::Core::Http::Request request(Azure::Core::Http::HttpMethod::Get, url);
 
     std::unique_ptr<Azure::Core::Http::RawResponse> response;
-    EXPECT_NO_THROW(response = pipeline.Send(Azure::Core::GetApplicationContext(), request));
-    auto responseCode = response->GetStatusCode();
-    int expectedCode = 200;
-    EXPECT_PRED2(
+    EXPECT_NO_THROW(response = pipeline.Send(Azure::Core::Context::GetApplicationContext(),
+  request)); auto responseCode = response->GetStatusCode(); int expectedCode = 200; EXPECT_PRED2(
         [](int expected, int code) { return expected == code; },
         expectedCode,
         static_cast<typename std::underlying_type<Azure::Core::Http::HttpStatusCode>::type>(
@@ -192,7 +188,8 @@ namespace Azure { namespace Core { namespace Test {
     Azure::Core::Http::Request request(Azure::Core::Http::HttpMethod::Get, url);
 
     std::unique_ptr<Azure::Core::Http::RawResponse> response;
-    EXPECT_NO_THROW(response = pipeline.Send(request, Azure::Core::GetApplicationContext()));
+    EXPECT_NO_THROW(
+        response = pipeline.Send(request, Azure::Core::Context::GetApplicationContext()));
     auto responseCode = response->GetStatusCode();
     int expectedCode = 200;
     EXPECT_PRED2(
@@ -222,7 +219,8 @@ namespace Azure { namespace Core { namespace Test {
     Azure::Core::Http::Request request(Azure::Core::Http::HttpMethod::Get, url);
 
     std::unique_ptr<Azure::Core::Http::RawResponse> response;
-    EXPECT_NO_THROW(response = pipeline.Send(request, Azure::Core::GetApplicationContext()));
+    EXPECT_NO_THROW(
+        response = pipeline.Send(request, Azure::Core::Context::GetApplicationContext()));
     auto responseCode = response->GetStatusCode();
     int expectedCode = 200;
     EXPECT_PRED2(
@@ -257,7 +255,8 @@ namespace Azure { namespace Core { namespace Test {
       Azure::Core::Http::Request request(Azure::Core::Http::HttpMethod::Get, url);
 
       std::unique_ptr<Azure::Core::Http::RawResponse> response;
-      EXPECT_NO_THROW(response = pipeline.Send(request, Azure::Core::GetApplicationContext()));
+      EXPECT_NO_THROW(
+          response = pipeline.Send(request, Azure::Core::Context::GetApplicationContext()));
       auto responseCode = response->GetStatusCode();
       int expectedCode = 200;
       EXPECT_PRED2(