+++ sdk/core/azure-core/CHANGELOG.md
@@ -15,7 +15,7 @@
 - Renamed `NoRevoke` to `EnableCertificateRevocationListCheck` for `Azure::Core::Http::CurlTransportSSLOptions`.
 - Renamed `GetString()` to `ToString()` in `Azure::Core::DateTime`.
 - Renamed `GetUuidString()` to `ToString()` in `Azure::Core::Uuid`.
-- Moved `Version` from `Azure::Core::Details` to `Azure::Core`.
+- Renamed `Azure::Core::Details::Version` to `Azure::Core::PackageVersion`.
 - Moved `BodyStream` and its derived types from `Azure::Core::Http` namespace to `Azure::IO`, and moved the `body_stream.hpp` header from `azure/core/http` to `azure/core/io`.
 - Moved `NullBodyStream` to internal usage only. It is not meant for public use.
 - Removed `LimitBodyStream`.
@@ -39,6 +39,7 @@
 - Moved `Azure::Core::Response<T>` to `Azure::Response<T>`.
 - Moved types in the `Azure::IO` namespace like `BodyStream` to `Azure::Core::IO`.
 - Moved `Azure::Core::ETag` to `Azure::ETag`.
+- Moved `Azure::Core::DateTime` to `Azure::DateTime`.
 
 ### Bug Fixes